Modify Rbum kernel model (in progress)

+++ basic/src/rbum/helper/rbum_scope_helper.rs
@@ -1,6 +1,6 @@
 use itertools::Itertools;
 
-pub fn get_pre_paths(scope_level: i32, own_paths: &str) -> String {
+pub fn get_pre_paths(scope_level: u8, own_paths: &str) -> String {
     let own_paths = own_paths.trim();
     let own_paths = own_paths.strip_suffix('/').unwrap_or(own_paths).to_string();
     if scope_level == 0 || own_paths.is_empty() {
@@ -16,7 +16,7 @@ pub fn get_pre_paths(scope_level: i32, own_paths: &str) -> String {
     split_items.iter().take(scope_level as usize).join("/")
 }
 
-pub fn get_path_item(scope_level: i32, own_paths: &str) -> Option<String> {
+pub fn get_path_item(scope_level: u8, own_paths: &str) -> Option<String> {
     let own_paths = own_paths.trim();
     let own_paths = own_paths.strip_suffix('/').unwrap_or(own_paths).to_string();
     if scope_level == 0 || own_paths.is_empty() {
